^&%^*$%^ FINALLY!!!
The key is to define tokens correctly (and with the correct priority)

because that helps writing the grammar and the parser.
The lexer will greedily seek to match tokens. When the priority of tokens is correct
(in our case, colons2xplus - colon - any_but_colon) the parser is not be subject to
ambiguity when matching and firing rules. On the other hand, if a grammar rule contains
a terminal (e.g. literal ':::'), but such terminal may be captured by the parser as
different combinations of tokens (e.g. %colon %colon %colon, or %colon2x %colon),
the parser will either be confused (and enter ambiguity) or would simply
fail because there is no grammar rule that matches (e.g. "::" _ ":":+).

The key was to define token %colons2xplus and give it the highest
priority, followed by %colon token (to match single colon), and finally
%any_but_colon to match everything else (one character at the time,
including newline).